apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "voicecall-insforge-refunds"
version = "0.1.0"
description = "FastAPI refund-call demo with Insforge-backed ecommerce tables"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "PyJWT>=2.10.0",
  "uvicorn[standard]>=0.30.0",
]

[tool.setuptools]
packages = ["app", "app.services"]
